Reveal file in the middle of Project Navigator (#1352)
* feat: reveal file in the middle of project navigator if it is out of bound

* feat: call scrollRowToVisible before moving to center

* feat: add reveal file action in navigate command

* feat: reveal the file forcefully if setting is off

* feat: enable reveal button

* chore: fix swiftlint error
